+++ LotCoMClient/App.xaml.cs
@@ -1,14 +1,23 @@
-﻿namespace LotCoMClient;
+﻿using LotCoMClient.Views;
 
-public partial class App : Application
-{
-	public App()
-	{
+namespace LotCoMClient;
+
+public partial class App : Application {
+	public App() {
 		InitializeComponent();
 	}
 
-	protected override Window CreateWindow(IActivationState? activationState)
-	{
-		return new Window(new AppShell());
+	protected override Window CreateWindow(IActivationState? activationState) {
+		// create the Main Window
+		Window MainWindow = new Window(new NavigationPage(new HomePage("Welcome")));
+		// create a TitleBar for the Window
+		TitleBar MainWindowTitleBar = new TitleBar {
+			Icon = "lotcom_logo.png",
+			Title = "LotCom Client",
+			Subtitle = "v0.1.0.0-alpha"      
+		};
+		// add the TitleBar to the Main Window
+		MainWindow.TitleBar = MainWindowTitleBar;
+		return MainWindow;
 	}
 }