+++ LotCoMClient/Views/StyleTestPage.xaml.cs
@@ -1,0 +1,7 @@
+namespace LotCoMClient.Views;
+
+public partial class StyleTestPage : ContentPage {
+    public StyleTestPage() {
+        InitializeComponent();
+    }
+}